Examples refactor and updates Updates to README
