+++ examples/observe/client/main.go
