+++ examples/observe/server/main.go
