+++ examples/proxy/coap/client/main.go
