+++ examples/proxy/coap/server/main.go
